refactor: add common types to separate files
