+++ packages/reports/src/Types/index.ts
@@ -1,0 +1,1 @@
+export * from './common-prop.type';
